requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
beautifulsoup4==4.6.0
chardet==3.0.4            # via readability-lxml
click==6.7                # via flask, python-dotenv
cssselect==1.0.1          # via readability-lxml
elasticsearch==5.4.0
flask-cors==3.0.3
flask==0.12.2
gunicorn==19.7.1
html2text==2016.9.19
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
lxml==3.8.0               # via readability-lxml
markupsafe==1.0           # via jinja2
mecab-python3==0.7
numpy==1.13.1             # via pandas
pandas==0.20.3
python-dateutil==2.6.1    # via pandas
python-dotenv==0.7.1
pytz==2017.2              # via pandas
readability-lxml==0.6.2
six==1.10.0               # via flask-cors, python-dateutil
unidecode==0.4.21
urllib3==1.22             # via elasticsearch
werkzeug==0.12.2          # via flask
